% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Adds two supplied numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{An integer or float}

\item{y}{An integer or float}
}
\value{
An integer or float that is the sum of the two supplied parameters
}
\description{
Adds two supplied numbers
}
\examples{
add(1,2)
add(10.65, 20.78)
}
