% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtracts the second parameter from the first}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{An integer or float}

\item{y}{An integer or float}
}
\value{
The difference between x and y, and integer or float
}
\description{
Subtracts the second parameter from the first
}
\examples{
subtract(1,2)
subtract(10.87, 1.111)
}
